Fix adjustments
Start datetime is simply defaulted to now.

End datetime defaults to start datetime, and then is adjusted to one hour after if possible.

Then, if end datetime is in the past, check if we can adjust both dates. If so, adjust them till both are in the future.

Also fixed spelling errors.

Signed-off-by: Shen Yichen <2007.yichen@gmail.com>